opt/apps/server-a/code/auth-backend/requirements.txt
Django>=4.2,<5.0
mongoengine>=0.28,<0.29
djangorestframework>=3.14,<4.0
django-cors-headers>=4.4,<5.0
PyJWT>=2.8,<3.0
pymongo>=4.6,<4.7
requests>=2.31,<3.0
gunicorn>=21.2,<22.0
python-dateutil>=2.8,<3.0
mongomock>=4.1,<5.0
